pyproject.toml
[project]
name = "productive-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=2.0.3",
    "openai>=1.107.1",
    "python-dotenv>=1.1.1",
    "tavily-python>=0.7.12",
    "yfinance>=0.2.65",
    "requests>=2.31.0",
]
